requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Database
chromadb==0.4.18
numpy==1.24.3
faiss-cpu==1.7.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
fitz==0.0.1.dev2  # PyMuPDF for advanced PDF processing
pymupdf==1.23.9

# Image Processing & OCR
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0

# Table Extraction
tabula-py==2.8.2
pandas==2.1.3

# Embeddings and NLP
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
tokenizers==0.15.0
huggingface-hub==0.19.4

# Text Processing
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
rank-bm25==0.2.2

# HTTP Client for Ollama
httpx==0.25.2
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI
factory-boy==3.3.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# File Processing
python-magic==0.4.27
filetype==1.2.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Evaluation & Metrics
seqeval==1.2.2
rouge-score==0.1.2
bert-score==0.3.13 